+++ R/guess_dts.R
@@ -1,17 +1,23 @@
-# - called by dts accessors, it attribute is not yet present.
-# - can assume x is dts, not test needed
-# - same name as dts accessor
-# - should never manipulate x
-
-
-guess_tattr <- function(x){
+#' Guess Time Attribute
+#'
+#' @param x a 'dts'
+#'
+#' - called by dts accessors, it attribute is not yet present.
+#' - can assume x is dts, not test needed
+#' - same name as dts accessor
+#' - should never manipulate x
+#'
+#' @examples
+#' guess_tattr(ts_dts(mdeaths))
+#' @noRd
+guess_tattr <- function(x) {
   x.time <- x[[dts_cname(x)$time]]
   class <- class(x.time)[1]
   if (!(class %in% c("Date", "POSIXct"))) {
-    stop("[time] col is not of class 'Date' or 'POSIXct'", call. = FALSE)
+    stop0("[time] column is not of class 'Date' or 'POSIXct'")
   }
-  if (class == "POSIXct"){
-    tz <- attr(x.time, 'tzone')
+  if (identical(class, "POSIXct")) {
+    tz <- attr(x.time, "tzone")
   } else {
     tz <- ""
   }
@@ -21,6 +27,19 @@ guess_tattr <- function(x){
   )
 }
 
+
+#' Guess Column Names
+#'
+#' @param x a 'dts'
+#'
+#' - called by dts accessors, it attribute is not yet present.
+#' - can assume x is dts, not test needed
+#' - same name as dts accessor
+#' - should never manipulate x
+#'
+#' @examples
+#' guess_cname(ts_dts(mdeaths))
+#' @noRd
 guess_cname <- function(x) {
   value.name <- guess_value(x)
   time.name <- guess_time(x, value.name = value.name)
@@ -31,20 +50,16 @@ guess_cname <- function(x) {
   }
   if (value.name != "value") {
     msg <- paste0(msg, "[value]: '", value.name, "' ")
-    # check if data frame is incidentally wide
-    cnames <- colnames(x)
-    idx.time <- which(cnames == time.name)
-    if ((idx.time - length(cnames)) > 1){
-      cols.r.of.time <- cnames[(idx.time + 1):length(cnames)]
-      value.cols <- vapply(x[, cols.r.of.time, with = FALSE], is_value, TRUE)
-      if (sum(value.cols) > 1){
-        message(
-        "More than one value column detected after the time colum, using the",
-        "outermost.\n",
-        "Are you using a wide data frame? ",
-        "To convert, use 'ts_long'.\n"
-        )
-      }
+    # check if data frame is incidentally wide (numeric id columns)
+    non_value <- setdiff(colnames(x), value.name)
+    numeric.id.cols <- vapply(x[, non_value, with = FALSE], is.numeric, TRUE)
+    if (sum(numeric.id.cols) > 0) {
+      message(
+        "Found numeric [id] column(s): ",
+        paste_quoted(names(numeric.id.cols)[numeric.id.cols]),
+        ".\nAre you using a wide data frame? To convert, use 'ts_long()'.",
+        "\nConvert column(s) to character or factor to silence this message.\n"
+      )
     }
   }