+++ R/to_from_all.R
@@ -1,21 +1,34 @@
-# helper functions for converters
-# (need to be adjusted if new classes are added)
-
-
-
 .tsbox_registry <- new.env(parent = emptyenv())
 
-register_class <- function(tsbox.class, actual.class = tsbox.class){
+#' Register a Time Series Class
+#'
+#' So tsbox is aware of it. Run when tsbox is loaded.
+#'
+#' @param tsbox.class character, class name, as used by tsbox
+#' @param actual.class character, the actual class name, which may be different.
+#'   E.g., tsbox calls tibbles `tbl`, so it can convert time series through
+#'   `ts_tbl()`, while their actual name is `"tbl_df"`.
+#' @noRd
+register_class <- function(tsbox.class, actual.class = tsbox.class) {
   class <- setNames(actual.class, tsbox.class)
   classes <- c(class, .tsbox_registry$class)
-  # to keep names
-  classes <- classes[classes == unique(classes)]
+  # unique() that keep names
+  classes <- classes[unique(names(classes))]
   assign("class", classes, envir = .tsbox_registry)
 }
 
+
+# register dts class
 register_class("dts")
 
-supported_classes <- function(){
+
+#' Supported Classes
+#'
+#' Returns all classes that have been registered.
+#' @examples
+#' supported_classes()
+#' @noRd
+supported_classes <- function() {
   .tsbox_registry$class
 }
 
@@ -24,13 +37,15 @@ supported_classes <- function(){
 #'
 #' Mainly used internally.
 #'
-#' @inherit ts_dts
+#' @inherit ts_default
+#' @return character, the relevant class of ts-boxable object
 #' @examples
 #' relevant_class(AirPassengers)
-#' relevant_class(ts_df(AirPassengers))
+#' x <- ts_df(AirPassengers)
+#' relevant_class(x)
 #' @export
 relevant_class <- function(x) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
   intersect(class(x), supported_classes())[1]
 }
 
@@ -39,8 +54,9 @@ relevant_class <- function(x) {
 #'
 #' Mainly used internally.
 #'
-#' @inherit ts_dts
-#' @return logical, either `TRUE` or `FALSE`
+#' @inherit ts_default
+#' @return logical, either `TRUE` or `FALSE`. `check_ts_boxable()` fails if not
+#'   `TRUE`
 #' @examples
 #' ts_boxable(AirPassengers)
 #' ts_boxable(lm)
@@ -49,11 +65,28 @@ ts_boxable <- function(x) {
   any(supported_classes() %in% class(x))
 }
 
+#' Error Check Functions
+#'
+#' @inherit ts_default
+#'
+#' @name ts_boxable
+#' @export
+check_ts_boxable <- function(x) {
+  if (!ts_boxable(x)) {
+    stop0(
+      "object is of non-ts-boxable class(es) ",
+      paste(paste0("'", class(x), "'"), collapse = ", "),
+      ". See `?ts_ts`."
+    )
+  }
+}
 
-# Universal Converter Function
-#
-# @param x time series object, either `ts`, `xts`, `data.frame` or `data.table`.
-# @return returns a function
+
+#' Universal Converter Function
+#'
+#' @param x time series object, either `ts`, `xts`, `data.frame` or `data.table`.
+#' @return returns a function
+#' @noRd
 as_class <- function(x) {
   stopifnot(any(supported_classes() %in% x))
 
@@ -62,10 +95,19 @@ as_class <- function(x) {
 }
 
 
-
+#' Desired Class
+#'
+#' If different time series objects are combined in `ts_c()`, `ts_bind()` etc.,
+#' which class should be returned
+#'
+#' @param ll a list with time series objects, usually collected via list(...)
+#' @return character string
+#' @examples
+#' desired_class(list(mdeaths, ts_df(fdeaths)))
+#' @noRd
 desired_class <- function(ll) {
   z <- unique(vapply(ll, relevant_class, ""))
-  if (length(z) == 1) {
+  if (length(z) == 1L) {
     if (z == "ts") {
       # no "ts" if mixed high frequecies
       uf <- unique(vapply(ll, frequency, 1))
@@ -81,7 +123,6 @@ desired_class <- function(ll) {
 }
 
 
-
 #' Re-Class ts-Boxable Object
 #'
 #' Copies class attributes from an existing ts-boxable series. Mainly used
@@ -89,22 +130,30 @@ desired_class <- function(ll) {
 #'
 #' Inspired by `xts::reclass`, which does something similar.
 #'
-#' @inherit ts_dts
+#' @inherit ts_default
 #' @param template ts-boxable time series, an object of class `ts`, `xts`,
-#'   `zoo`, `data.frame`, `data.table`, `tbl`, `tbl_ts`, `tbl_time`, `tis`,
-#'   `irts` or `timeSeries`.
+#'   `zoo`, `zooreg`, `data.frame`, `data.table`, `tbl`, `tbl_ts`, `tbl_time`,
+#'   `tis`, `irts` or `timeSeries`.
 #' @param preserve.mode should the mode the time column be preserved
 #'   (data frame only)
 #' @param preserve.names should the name of the time column be preserved
 #'   (data frame only)
 #' @param preserve.time should the values time column be preserved
 #'   (data frame only)
+#' @param preserve.attr should the attributes of the value column be preserved
+#'   (data frame only)
+#' @return a ts-boxable object of the same class as `template`,
+#'   i.e., an object of class `ts`, `xts`, `zoo`,
+#'   `data.frame`, `data.table`, `tbl`, `tbl_ts`, `tbl_time`, `tis`, `irts` or
+#'   `timeSeries`.
 #' @export
+#' @examples
+#' copy_class(mdeaths, ts_tbl(fdeaths))
 copy_class <- function(x, template,
                        preserve.mode = TRUE,
                        preserve.names = FALSE,
-                       preserve.time = FALSE) {
-
+                       preserve.time = FALSE,
+                       preserve.attr = TRUE) {
   if (!ts_boxable(x)) {
     if (inherits(template, "ts")) {
       x <- ts(x)
@@ -120,13 +169,13 @@ copy_class <- function(x, template,
   }
 
   # to deal with 1 period time series: separate ts treatment
-  if (inherits(template, "ts")){
-    if (inherits(x, "ts")){
+  if (inherits(template, "ts")) {
+    if (inherits(x, "ts")) {
       ans <- x
     } else {
       x.dts <- ts_dts(x)
       # is there only one observation?
-      if (number_of_series(x.dts) == nrow(ts_dts(x.dts))){
+      if (number_of_series(x.dts) == nrow(ts_dts(x.dts))) {
         ans <- ts_ts_dts(ts_dts(x), frequency = frequency(template))
       } else {
         ans <- as_class(relevant_class(template))(x)
@@ -152,12 +201,16 @@ copy_class <- function(x, template,
       ans[[tn]] <- template[[tn]]
     }
 
+    if (preserve.attr) {
+      vn <- guess_value(ans)
+      attributes(ans[[vn]]) <- attributes(template[[vn]])
+    }
+
     if (preserve.names) {
-      if (!identical(names(ans), names(template))){
+      if (!identical(names(ans), names(template))) {
         names(ans) <- names(template)
       }
     }
-
   }
 
   ans