+++ R/to_from_tibble.R
@@ -2,11 +2,15 @@ register_class("tbl", "tbl_df")
 
 # to ---------------------------------------------------------------------------
 
+#' Convert to Class
+#' @noRd
 ts_tbl_dts <- function(x) {
   stopifnot(requireNamespace("tibble"))
   tibble::as_tibble(as.data.frame(ts_data.table(x)))
 }
 
+#' Convert to Class
+#' @noRd
 as.tbl_df <- function(x) {
   stopifnot(requireNamespace("tibble"))
   tibble::as_tibble(x)
@@ -22,7 +26,9 @@ as.tbl_df <- function(x) {
 #' @name ts_ts
 #' @export
 ts_tbl <- function(x) {
-  stopifnot(ts_boxable(x))
-  if (relevant_class(x) == "tbl") return(x)
+  check_ts_boxable(x)
+  if (relevant_class(x) == "tbl") {
+    return(x)
+  }
   ts_tbl_dts(ts_dts(x))
 }