+++ R/to_from_tslist.R
@@ -2,10 +2,12 @@ register_class("tslist")
 
 # to ---------------------------------------------------------------------------
 
+#' Convert to Class
+#' @noRd
 ts_tslist_dts <- function(x) {
   stopifnot(inherits(x, "dts"))
   x <- combine_id_cols(x)
-  if (number_of_series(x) == 1) {
+  if (number_of_series(x) == 1L) {
     z <- list(ts_ts(x))
     # if a single series has an id, use to name element
     cid <- dts_cname(x)$id
@@ -43,7 +45,9 @@ ts_dts.tslist <- function(x) {
 #' @name ts_ts
 #' @export
 ts_tslist <- function(x) {
-  stopifnot(ts_boxable(x))
-  if (relevant_class(x) == "tslist") return(x)
+  check_ts_boxable(x)
+  if (relevant_class(x) == "tslist") {
+    return(x)
+  }
   ts_tslist_dts(ts_dts(x))
 }