+++ NEWS.md
@@ -1,3 +1,26 @@
+# tsbox 0.4  (2023-03-08)
+
+## Features
+
+- Comply with rOpenSci Standards
+- Better test coverage
+- Support for 'zooreg' time series
+
+## Documentation
+
+- better error messages when input is not ts-boxable
+- better error messages when data is non-standard
+- New repository: https://github.com/ropensci
+- New documentation website: https://docs.ropensci.org/tsbox
+
+## Bug fixes
+
+- ts_long() works with on a wider set of data frames
+- ts_plot() works better when applied on non-standard data frame
+- ts_first_of_period works with POSIXct #210
+- avoid date classes not getting respected by stopifnot, #213
+
+
 # tsbox 0.3.1 (2021-09-16)
 
 ## Bug fixes